Migrate to null safety

+++ lib/src/input_formatters/spaced_text_input_formatter.dart
@@ -24,7 +24,7 @@ class SpacedTextInputFormatter extends TextInputFormatter {
     return TextEditingValue(
       text: splitText,
       selection: TextSelection.collapsed(
-        offset: newOffset,
+        offset: newOffset as int,
       ),
     );
   }
